feat: add recipients avatars on all tables
