+++ apps/web/src/components/(dashboard)/avatar/stack-avatar.tsx
