+++ packages/lib/types/document-with-recipient.ts
@@ -1,0 +1,5 @@
+import { Document, Recipient } from '@documenso/prisma/client';
+
+export type DocumentWithReciepient = Document & {
+  Recipient: Recipient[];
+};
